# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Portal API",
        "description": "API for News Portal",
        "contact": {
            "name": "Your Name",
            "email": "your@email.com"
        },
        "license": {
            "name": "Your License",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/paginated": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Search for news articles",
                "description": "Searches for news articles based on the provided parameters.",
                "operationId": "searchAllNews",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "The search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort option for the results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "relevancy",
                                "popularity",
                                "publishedAt"
                            ]
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language code for filtering results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "publishedAt": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fetch-searched-news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Fetch searched news articles",
                "description": "Fetches news articles based on the provided search parameters.",
                "operationId": "fetchSearchedNews",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "The search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort option for the results",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "relevancy",
                                "popularity",
                                "publishedAt"
                            ]
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language code for filtering results",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of articles per page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "publishedAt": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/paginated": {
            "get": {
                "tags": [
                    "Top News"
                ],
                "summary": "Get top news",
                "description": "Fetches top news based on the specified category and country.",
                "operationId": "getTopNews",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "The category of news (business, entertainment, health, science, sports, technology).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "business",
                                "entertainment",
                                "health",
                                "science",
                                "sports",
                                "technology"
                            ]
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "The country code for news filtering (e.g., 'us', 'gb', 'ca').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "publishedAt": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}